11 - move settings and buttons
