--- src/elementID.js
@@ -1,3 +1,0 @@
-export const ElementID = {
-    SEARCH_BAR_ID: "SearchBar"
-    };
