+++ src/identifiers.js
@@ -1,0 +1,14 @@
+export const ElementID = {
+    SEARCH_BAR_ID: "SearchBar"
+    };
+
+
+export const SettingsID = {
+    PANEL_ID: "settings-panel",
+    COLOR_THEME: "color-theme",
+    HOVER_COLOR: "hover-color-theme",
+    COMPACT_CARDS: "compact-data-cards",
+    ANIMATIONS: "animations-switch",
+    BUTTON_LABELS: "button-labels",
+    UPDATE_SETTINGS: "update-settings-button"
+    };
