+++ src/search/index.js
@@ -1,0 +1,3 @@
+export * as models from './models.js';
+export * as functions from './functions.js';
+export * as view from './view.js';
