+++ src/search/view.js
@@ -1,4 +1,4 @@
-import { ElementID } from "../elementID.js";
+import { ElementID } from "../identifiers.js";
 
 export function resetSearchBar(numberOfCards) {
     let searchBar = document.getElementById(ElementID.SEARCH_BAR_ID);
