+++ src/tree/enums.js
@@ -1,1 +1,5 @@
-// add enum for tree depth
+export const TreeDepth = {
+    ROOT: 0,
+    BRANCH: 1,
+    LEAF: 2
+    };
